[WIP] started refactoring
